+++ config/config/dep.go
@@ -1,0 +1,6 @@
+package config
+
+type Dep struct {
+	Name string
+	Args []string
+}